services/recommendation-service/requirements.txt
kafka-python==2.0.2
redis==4.3.4
structlog==22.3.0